+++ src/routes/indexRouter.js
@@ -4,11 +4,14 @@ const indexRouter = Router()
 
 indexRouter
   .get('/',(req, res)=>{
-    res.send('<h1>Dev Jobs - Home Page</h1>')
+    //(V)
+    res.render('home')
   })
 
-indexRouter.get('/about',(req, res)=>{
-  res.send('<h1>About Page</h1>')
-})
+indexRouter
+  .get('/about',(req, res)=>{
+    //(V)
+    res.render('about', {info: 'We like to code here'})
+  })
 
 module.exports = indexRouter